SalesState alpha
I have the buttons for the class mostly added. However, I can't figure
out how to do the add product method in general. If someone could help
me out on getting that done, that would be great. I do have it set up
where you can access the add product method, and click the back button
to get back to the main salesclerk GUI, but that's about as far as I
got.